refactor: PayInfoProvider 파일 위치 domain 폴더 안으로 변경
